Refactored models.py into a directory of separate models
