+++ alayatodo/models/__init__.py
@@ -1,0 +1,2 @@
+from .Todo import Todo
+from .User import User
